fix: type mismatch across types

+++ app/composeApp/src/commonMain/kotlin/com/altmanifest/secondtake/domain/Rating.kt
@@ -5,6 +5,6 @@ import kotlin.time.Duration
 
 data class Rating(val value: Double, val age: Duration) {
     fun pointDifference(other: Rating): Double = abs(other.value - this.value)
-    fun comparableWith(other: Rating, tolerance: Int): Boolean = pointDifference(other) <= tolerance
-    fun notComparableWith(other: Rating, tolerance: Int): Boolean = !comparableWith(other, tolerance)
+    fun comparableWith(other: Rating, tolerance: Double): Boolean = pointDifference(other) <= tolerance
+    fun notComparableWith(other: Rating, tolerance: Double): Boolean = !comparableWith(other, tolerance)
 }
